fleet-adt4: split tui into screens/ package (option 2) + tests
