+++ dcist_launch_system/src/dcist_launch_system/tui/__init__.py
@@ -562,2550 +562,12 @@ def zenoh(ctx, machine, dry_run):
 
 
 # ---------------------------------------------------------------------------
-# TUI mode (textual)
+
+# ---------------------------------------------------------------------------
+# TUI mode (textual) — delegated to dcist_launch_system.tui.app
 # ---------------------------------------------------------------------------
 
-
-def _launch_tui(network, topology_path, output_root):
-    """Launch the interactive Textual TUI."""
-    try:
-        from textual.app import App, ComposeResult
-        from textual.binding import Binding
-        from textual.containers import Horizontal, Vertical, VerticalScroll
-        from textual.screen import ModalScreen
-        from textual.widgets import (
-            Button,
-            DataTable,
-            Footer,
-            Header,
-            Input,
-            Label,
-            ProgressBar,
-            RichLog,
-            Rule,
-            SelectionList,
-            Static,
-            Tree,
-        )
-    except ImportError:
-        click.secho(
-            "textual not installed. Install with: pip install textual\n"
-            "Falling back to CLI mode. Use subcommands (status, maps, etc.).",
-            fg="yellow",
-        )
-        return
-
-    topo = load_topology(topology_path)
-    active_network = network or detect_network(topo)
-    if not active_network:
-        click.secho("Could not detect network. Use --network.", fg="red")
-        sys.exit(1)
-
-    all_machines = resolve_machines(topo, active_network)
-
-    # Runtime config overrides (editable in TUI)
-    runtime_config = {m["name"]: dict(m) for m in all_machines}
-    _deleted_machines = set()  # track explicit removals for save
-
-    # Cache local IPs for detecting when a base station is the local machine
-    from dcist_launch_system.fleet_helpers import get_local_ips
-    _local_ips = set(get_local_ips())
-
-    def _fuzzy_match(query, text):
-        """Subsequence fuzzy match — all chars of query appear in order in text."""
-        if not query:
-            return True
-        query = query.lower()
-        text = text.lower()
-        qi = 0
-        for ch in text:
-            if ch == query[qi]:
-                qi += 1
-                if qi == len(query):
-                    return True
-        return False
-
-    # ---- Maps Screen ----
-
-    class ConfirmScreen(ModalScreen[bool]):
-        BINDINGS = [
-            Binding("y", "confirm_yes", "Yes", priority=True),
-            Binding("n", "confirm_no", "No", priority=True),
-            Binding("escape", "confirm_no", "Cancel", priority=True),
-        ]
-
-        def __init__(self, message: str, yes_label: str = "Yes, Proceed (y)", no_label: str = "No, Cancel (n)"):
-            super().__init__()
-            self.message = message
-            self._yes_label = yes_label
-            self._no_label = no_label
-
-        def compose(self) -> ComposeResult:
-            yield Vertical(
-                Label(self.message),
-                Rule(),
-                Label("[dim]y=yes  n/escape=no[/]"),
-                Horizontal(
-                    Button(self._yes_label, variant="error", id="btn_yes"),
-                    Button(self._no_label, variant="primary", id="btn_no"),
-                ),
-                id="confirm_kill_dialog"
-            )
-
-        def on_button_pressed(self, event: Button.Pressed) -> None:
-            self.dismiss(event.button.id == "btn_yes")
-
-        def action_confirm_yes(self):
-            self.dismiss(True)
-
-        def action_confirm_no(self):
-            self.dismiss(False)
-
-    def _check_sessions_and_prompt(app, targets, session_name, log):
-        """Check for running tmux sessions and prompt user in TUI threads to kill them."""
-        def check(m):
-            rc, _, _ = ssh_cmd(m["user"], m["ip"], f"tmux has-session -t {shlex.quote(session_name)} 2>/dev/null", timeout=5)
-            if rc == 0:
-                return m["name"]
-            return None
-
-        app.call_from_thread(log.write, f"  [dim]Checking for existing '{session_name}' sessions...[/]")
-        results = run_parallel(check, targets)
-        conflicts = [r for _, r in results if isinstance(r, str)]
-
-        if not conflicts:
-            return True
-
-        app.call_from_thread(
-            log.write,
-            f"  [yellow]Session '{session_name}' already running on: {', '.join(conflicts)}[/]"
-        )
-
-        user_decision = [None]
-        ev = threading.Event()
-
-        def ask():
-            def on_dismiss(result):
-                user_decision[0] = result
-                ev.set()
-            msg = f"Session '{session_name}' is already running on {len(conflicts)} machine(s):\n{', '.join(conflicts)}\n\nTerminate existing sessions and proceed?"
-            app.push_screen(ConfirmScreen(msg, yes_label="Yes, Terminate (y)", no_label="No, Abort (n)"), on_dismiss)
-
-        app.call_from_thread(ask)
-        ev.wait()
-
-        if user_decision[0]:
-            for m in targets:
-                if m["name"] in conflicts:
-                    app.call_from_thread(log.write, f"  [dim]Terminating session {session_name} on {m['name']}...[/]")
-                    ssh_cmd(m["user"], m["ip"], f"tmux kill-session -t {shlex.quote(session_name)}", timeout=5)
-            return True
-        else:
-            app.call_from_thread(log.write, "[red]Launch aborted by user.[/]")
-            return False
-
-    class PriorMapSelector(Vertical):
-        """Reusable widget for browsing and selecting prior maps from fleet nodes.
-
-        Uses the same Tree-based approach as MapsScreen. Click an experiment
-        to select it. Experiments with both hydra + roman are marked with a green check.
-        """
-        def compose(self) -> ComposeResult:
-            yield Label("Prior map (click to select, or type custom path):")
-            yield Input(placeholder="Filter experiments...", id="prior_search")
-            yield Tree("Fleet Experiments", id="prior_tree")
-            yield Input(placeholder="(optional) custom: machine:~/adt4_output/exp", id="prior_input")
-
-        def on_mount(self):
-            self._selected_path = ""
-            self._selected_source = None
-            self._all_results = []
-            self.set_timer(0.5, self._safe_refresh)
-
-        def _safe_refresh(self):
-            """Guard against NoActiveAppError on re-mount of cached screens."""
-            try:
-                self.app  # test if connected
-                self.refresh_priors()
-            except Exception:
-                pass
-
-        def refresh_priors(self, log_widget=None):
-
-            def do_fetch():
-                machines = [m for m in runtime_config.values() if m.get("online", False)]
-                def fetch(m):
-                    return list_remote_experiments(m["user"], m["ip"], output_root)
-                return run_parallel(fetch, machines)
-
-            def on_done(results):
-                self._all_results = results
-                try:
-                    self._populate_tree()
-                except Exception:
-                    pass
-                if log_widget:
-                    total = sum(
-                        len(exps) for _, exps in results
-                        if not isinstance(exps, Exception)
-                    )
-                    log_widget.write(f"[green]Found {total} experiments across fleet.[/]")
-
-            def _bg():
-                result = do_fetch()
-                try:
-                    self.app.call_from_thread(on_done, result)
-                except Exception:
-                    pass
-
-            threading.Thread(target=_bg, daemon=True).start()
-
-        def _populate_tree(self, query=""):
-            tree = self.query_one("#prior_tree", Tree)
-            tree.clear()
-            for m, exps in sorted(self._all_results, key=lambda x: x[0]["name"]):
-                if isinstance(exps, Exception):
-                    node = tree.root.add(m["name"], expand=True)
-                    node.add_leaf(f"[red]Error: {exps}[/]")
-                    continue
-                filtered = [
-                    e for e in sorted(exps, key=lambda e: e["name"])
-                    if _fuzzy_match(query, e["name"])
-                ] if query else sorted(exps, key=lambda e: e["name"])
-                if not filtered and query:
-                    continue
-                node = tree.root.add(m["name"], expand=True)
-                if not filtered:
-                    node.add_leaf("[dim]no experiments[/]")
-                    continue
-                for exp in filtered:
-                    subdirs = exp.get("subdirs", [])
-                    has_prior = "hydra" in subdirs and "roman" in subdirs
-                    tag = " [green]\u2714[/]" if has_prior else ""
-                    label = f"{exp['name']}  ({exp.get('size', '?')}){tag}"
-                    leaf = node.add_leaf(label)
-                    leaf.data = {
-                        "machine": m["name"],
-                        "experiment": exp["name"],
-                        "path": f"{output_root}/{exp['name']}",
-                    }
-            tree.root.expand_all()
-
-        def on_input_changed(self, event: Input.Changed):
-            if event.input.id == "prior_search" and self._all_results:
-                self._populate_tree(event.value.strip())
-
-        def on_tree_node_selected(self, event):
-            node = event.node
-            if hasattr(node, "data") and node.data and "experiment" in node.data:
-                self._selected_path = node.data["path"]
-                self._selected_source = node.data["machine"]
-
-        def get_prior_map(self):
-            """Return (prior_path, source_machine_name) or ("", None)."""
-            custom = self.query_one("#prior_input", Input).value.strip()
-            if custom:
-                if ":" in custom:
-                    parts = custom.split(":", 1)
-                    return parts[1], parts[0]
-                return custom, None
-            if self._selected_path:
-                return self._selected_path, self._selected_source
-            return "", None
-
-    class MapsScreen(ModalScreen):
-        BINDINGS = [
-            Binding("escape", "dismiss", "Back", priority=True),
-            Binding("t", "transfer_selected", "Transfer", priority=True),
-            Binding("v", "verify_selected", "Verify", priority=True),
-            Binding("r", "refresh_maps", "Refresh", priority=True),
-        ]
-
-        def __init__(self):
-            super().__init__()
-            self.selected_machine = None
-            self.selected_experiment = None
-            self._all_results = []  # cached fetch results
-
-        def compose(self) -> ComposeResult:
-            yield Vertical(
-                Label("[bold]Map Browser[/] — select an experiment"),
-                Input(placeholder="Search experiments...", id="map_search"),
-                Rule(),
-                Tree("Fleet Maps", id="map_tree"),
-                Rule(),
-                RichLog(id="map_log", highlight=True, markup=True, max_lines=50),
-            )
-            yield Footer()
-
-        def action_refresh_maps(self):
-            self._refresh_maps()
-
-        def on_mount(self):
-            self._refresh_maps()
-
-        def _refresh_maps(self):
-            tree = self.query_one("#map_tree", Tree)
-            log = self.query_one("#map_log", RichLog)
-            tree.clear()
-            log.write("[dim]Fetching experiments from fleet...[/]")
-
-            def do_fetch():
-                machines = [m for m in runtime_config.values() if m.get("online", False)]
-
-                def fetch(m):
-                    return list_remote_experiments(m["user"], m["ip"], output_root)
-
-                return run_parallel(fetch, machines)
-
-            def on_done(results):
-                self._all_results = results
-                query = self.query_one("#map_search", Input).value.strip()
-                self._populate_tree(query)
-                log.write("[green]Maps loaded.[/]")
-
-            def bg():
-                results = do_fetch()
-                self.app.call_from_thread(on_done, results)
-
-            threading.Thread(target=bg, daemon=True).start()
-
-        def _populate_tree(self, query=""):
-            tree = self.query_one("#map_tree", Tree)
-            tree.clear()
-            for m, exps in sorted(self._all_results, key=lambda x: x[0]["name"]):
-                if isinstance(exps, Exception):
-                    node = tree.root.add(m["name"], expand=True)
-                    node.data = {"machine": m["name"]}
-                    node.add_leaf(f"[red]Error: {exps}[/]")
-                    continue
-                filtered = [
-                    e for e in sorted(exps, key=lambda e: e["name"])
-                    if _fuzzy_match(query, e["name"])
-                ] if query else sorted(exps, key=lambda e: e["name"])
-                if not filtered and query:
-                    continue  # hide machines with no matches
-                node = tree.root.add(m["name"], expand=True)
-                node.data = {"machine": m["name"]}
-                if not filtered:
-                    node.add_leaf("[dim]no experiments[/]")
-                    continue
-                for exp in filtered:
-                    subdirs_str = ", ".join(exp["subdirs"]) if exp["subdirs"] else ""
-                    label = f"{exp['name']}  [{subdirs_str}]  ({exp['size']})"
-                    leaf = node.add_leaf(label)
-                    leaf.data = {"machine": m["name"], "experiment": exp["name"]}
-            tree.root.expand_all()
-
-        def on_input_changed(self, event: Input.Changed):
-            if event.input.id == "map_search" and self._all_results:
-                self._populate_tree(event.value.strip())
-
-        def on_tree_node_selected(self, event):
-            node = event.node
-            if hasattr(node, "data") and node.data and "experiment" in node.data:
-                self.selected_machine = node.data["machine"]
-                self.selected_experiment = node.data["experiment"]
-                log = self.query_one("#map_log", RichLog)
-                log.write(
-                    f"Selected: [bold]{self.selected_machine}[/]:"
-                    f"[cyan]{self.selected_experiment}[/]"
-                    f" — press [bold]T[/] to transfer, [bold]V[/] to verify"
-                )
-
-        def action_transfer_selected(self):
-            if not self.selected_experiment:
-                self.query_one("#map_log", RichLog).write("[yellow]Select an experiment first.[/]")
-                return
-            self.app.push_screen(
-                TransferScreen(self.selected_machine, self.selected_experiment)
-            )
-
-        def action_verify_selected(self):
-            if not self.selected_experiment:
-                self.query_one("#map_log", RichLog).write("[yellow]Select an experiment first.[/]")
-                return
-            self.app.push_screen(VerifyScreen(self.selected_experiment))
-
-    # ---- Transfer Screen ----
-
-    class TransferScreen(ModalScreen):
-        BINDINGS = [
-            Binding("escape", "dismiss", "Back", priority=True),
-            Binding("t", "start_transfer", "Start Transfer", priority=True),
-        ]
-
-        def __init__(self, src_machine, experiment):
-            super().__init__()
-            self.src_machine = src_machine
-            self.experiment = experiment
-            self._transferring = False
-
-        def compose(self) -> ComposeResult:
-            others = [
-                (m["name"], m["name"])
-                for m in runtime_config.values()
-                if m["name"] != self.src_machine and m.get("online", False)
-            ]
-            yield Vertical(
-                Label(
-                    f"[bold]Transfer[/] {self.src_machine}:{self.experiment} -> select destinations"
-                ),
-                Rule(),
-                SelectionList(*others, id="dest_select"),
-                Rule(),
-                Vertical(id="progress_area"),
-                RichLog(id="transfer_log", highlight=True, markup=True, max_lines=100),
-            )
-            yield Footer()
-
-        def action_start_transfer(self):
-            self._do_transfer()
-
-        def _do_transfer(self):
-            if self._transferring:
-                return
-            sel = self.query_one("#dest_select", SelectionList)
-            selected_names = list(sel.selected)
-            if not selected_names:
-                self.query_one("#transfer_log", RichLog).write(
-                    "[yellow]Select at least one destination.[/]"
-                )
-                return
-
-            self._transferring = True
-            log = self.query_one("#transfer_log", RichLog)
-            src = runtime_config[self.src_machine]
-            src_path = f"{output_root}/{self.experiment}"
-            dest_name = f"prior_{self.experiment}"
-
-            log.write(f"[yellow]Excluding rosbags. Saving as {dest_name}[/]")
-
-            # Build progress bars in the UI
-            area = self.query_one("#progress_area", Vertical)
-            area.remove_children()
-            for dname in selected_names:
-                area.mount(
-                    Vertical(
-                        Label(f"  {dname}:"),
-                        ProgressBar(total=100, show_eta=False, id=f"pb_{dname}"),
-                        id=f"pbrow_{dname}",
-                    )
-                )
-
-            def do_transfers():
-                try:
-                    for i, dname in enumerate(selected_names):
-                        dst = runtime_config[dname]
-                        dst_path = f"{output_root}/{dest_name}"
-                        self.app.call_from_thread(
-                            log.write, f"Transferring to [bold]{dname}[/]..."
-                        )
-
-                        def make_cb(name):
-                            def cb(pct):
-                                self.app.call_from_thread(
-                                    self._update_progress, name, pct
-                                )
-                            return cb
-
-                        ok, method, msg = rsync_transfer(
-                            src["user"], src["ip"], src_path,
-                            dst["user"], dst["ip"], dst_path,
-                            exclude=["recorded_data/"],
-                            progress_callback=make_cb(dname),
-                        )
-                        if ok:
-                            self.app.call_from_thread(
-                                self._update_progress, dname, 100
-                            )
-                            self.app.call_from_thread(
-                                log.write, f"  [green]{dname}: OK ({method})[/]"
-                            )
-                        else:
-                            self.app.call_from_thread(
-                                log.write, f"  [red]{dname}: FAIL ({method}) {msg}[/]"
-                            )
-                    self.app.call_from_thread(log.write, "[bold]Transfer complete.[/]")
-                finally:
-                    self._transferring = False
-
-            threading.Thread(target=do_transfers, daemon=True).start()
-
-        def _update_progress(self, dname, pct):
-            try:
-                bar = self.query_one(f"#pb_{dname}", ProgressBar)
-                bar.update(progress=pct)
-            except Exception:
-                pass
-
-    # ---- Verify Screen ----
-
-    class VerifyScreen(ModalScreen):
-        BINDINGS = [Binding("escape", "dismiss", "Back", priority=True)]
-
-        def __init__(self, experiment):
-            super().__init__()
-            self.experiment = experiment
-
-        def compose(self) -> ComposeResult:
-            yield Vertical(
-                Label(f"[bold]Verify:[/] {self.experiment}"),
-                Rule(),
-                DataTable(id="verify_table"),
-                RichLog(id="verify_log", highlight=True, markup=True, max_lines=50),
-                Footer(),
-            )
-
-        def on_mount(self):
-            log = self.query_one("#verify_log", RichLog)
-            log.write("[dim]Computing checksums across fleet...[/]")
-
-            def do_verify():
-                machines = [m for m in runtime_config.values() if m.get("online", False)]
-
-                def fetch(m):
-                    return hash_remote_experiment(m["user"], m["ip"], output_root, self.experiment)
-
-                return run_parallel(fetch, machines)
-
-            def on_done(results):
-                table = self.query_one("#verify_table", DataTable)
-                machine_hashes = {
-                    m["name"]: h for m, h in results if isinstance(h, dict) and h
-                }
-                if not machine_hashes:
-                    log.write("[yellow]No hash data returned.[/]")
-                    return
-
-                all_files = sorted(set(f for h in machine_hashes.values() for f in h))
-                mnames = sorted(machine_hashes.keys())
-
-                table.add_column("File", key="file")
-                for mn in mnames:
-                    table.add_column(mn, key=mn)
-
-                diffs = 0
-                for fpath in all_files:
-                    hashes = {mn: machine_hashes[mn].get(fpath) for mn in mnames}
-                    unique = set(v for v in hashes.values() if v)
-                    consistent = len(unique) <= 1 and all(v for v in hashes.values())
-                    row = [fpath]
-                    for mn in mnames:
-                        h = hashes[mn]
-                        if h is None:
-                            row.append("MISSING")
-                            consistent = False
-                        elif consistent:
-                            row.append("OK")
-                        else:
-                            row.append(h[:8])
-                    if not consistent:
-                        diffs += 1
-                    table.add_row(*row)
-
-                if diffs == 0:
-                    log.write("[bold green]All files consistent.[/]")
-                else:
-                    log.write(f"[bold red]{diffs} file(s) differ.[/]")
-
-            def bg():
-                results = do_verify()
-                self.app.call_from_thread(on_done, results)
-
-            threading.Thread(target=bg, daemon=True).start()
-
-    # ---- Config Screen ----
-
-    class EditMachineScreen(ModalScreen):
-        """Modal dialog for editing or adding a single machine entry."""
-        AUTO_FOCUS = "Input"
-        BINDINGS = [
-            Binding("escape", "cancel", "Cancel", priority=True),
-            Binding("f5", "save", "Save [F5]", priority=True),
-        ]
-        DEFAULT_CSS = """
-        EditMachineScreen { align: center middle; }
-        EditMachineScreen > Vertical {
-            width: 60;
-            height: auto;
-            border: thick $primary;
-            background: $surface;
-            padding: 1 2;
-        }
-        EditMachineScreen Input { width: 1fr; }
-        EditMachineScreen #err_label { color: $error; }
-        """
-
-        def __init__(self, name=None, minfo=None, existing_names=None, **kwargs):
-            super().__init__(**kwargs)
-            self._name = name
-            self._minfo = minfo or {}
-            self._existing_names = existing_names or set()
-
-        def compose(self) -> ComposeResult:
-            addrs = self._minfo.get("addresses", {})
-            title = f"Editing: {self._name}" if self._name else "Add New Machine"
-            rows = [Label(f"[bold]{title}[/]")]
-            sof = {"select_on_focus": False}
-            if not self._name:
-                rows.append(Input(placeholder="Name (e.g. newton)", id="ed_name", **sof))
-            rows += [
-                Input(value=self._minfo.get("role", "robot"), placeholder="Role (robot / base_station)", id="ed_role", **sof),
-                Input(value=self._minfo.get("platform_id", ""), placeholder="Platform (e.g. smaug)", id="ed_platform", **sof),
-                Input(value=addrs.get("mit_wifi", ""), placeholder="MIT WiFi IP (e.g. 10.29.x.x)", id="ed_mit_wifi", **sof),
-                Input(value=addrs.get("silvus", ""), placeholder="Silvus IP (e.g. 192.168.100.x)", id="ed_silvus", **sof),
-                Input(value=self._minfo.get("desc", ""), placeholder="Description", id="ed_desc", **sof),
-                Label("[dim]Tab=next field  F5=save  Escape=cancel[/]"),
-                Label("", id="err_label"),
-            ]
-            yield Vertical(*rows)
-
-        def action_cancel(self):
-            self.dismiss(None)
-
-        def action_save(self):
-            err = self.query_one("#err_label", Label)
-            if self._name:
-                name = self._name
-            else:
-                name = self.query_one("#ed_name", Input).value.strip()
-                if not name:
-                    err.update("[red]Name is required.[/]")
-                    return
-                if name in self._existing_names:
-                    err.update(f"[red]'{name}' already exists.[/]")
-                    return
-            role = self.query_one("#ed_role", Input).value.strip() or "robot"
-            if role not in ("robot", "base_station"):
-                err.update("[red]Role must be 'robot' or 'base_station'.[/]")
-                return
-            self.dismiss({
-                "name": name,
-                "role": role,
-                "platform": self.query_one("#ed_platform", Input).value.strip(),
-                "desc": self.query_one("#ed_desc", Input).value.strip(),
-                "mit_wifi": self.query_one("#ed_mit_wifi", Input).value.strip(),
-                "silvus": self.query_one("#ed_silvus", Input).value.strip(),
-            })
-
-    class ConfigScreen(ModalScreen):
-        """Topology editor — manages machines in topology.yaml."""
-        BINDINGS = [
-            Binding("escape", "dismiss", "Back", priority=True),
-            Binding("e", "edit_selected", "Edit [e]"),
-            Binding("enter", "edit_selected", "Edit", show=False),
-            Binding("f2", "add_machine_action", "Add [F2]", priority=True),
-            Binding("f3", "recheck_action", "Recheck [F3]", priority=True),
-            Binding("f4", "remove_selected", "Remove [F4]", priority=True),
-            Binding("f5", "save_action", "Save [F5]", priority=True),
-        ]
-
-        DEFAULT_CSS = """
-        ConfigScreen #config_table { height: auto; max-height: 16; }
-        """
-
-        def compose(self) -> ComposeResult:
-            yield VerticalScroll(
-                Label("[bold]Topology Editor[/]"),
-                Label("[dim]e=edit, F2=add, F4=remove, F3=recheck, F5=save to YAML[/]"),
-                Rule(),
-                DataTable(id="config_table"),
-                Rule(),
-                RichLog(id="config_log", highlight=True, markup=True, max_lines=20),
-            )
-            yield Footer()
-
-        def on_mount(self):
-            table = self.query_one("#config_table", DataTable)
-            table.add_column("", key="status")
-            table.add_column("Name", key="name")
-            table.add_column("Role", key="role")
-            table.add_column("Platform", key="platform")
-            table.add_column("MIT WiFi", key="mit_wifi")
-            table.add_column("Silvus", key="silvus")
-            table.add_column("User", key="user")
-            table.add_column("Desc", key="desc")
-            table.cursor_type = "row"
-            self._refresh_config_table()
-
-        def _refresh_config_table(self):
-            table = self.query_one("#config_table", DataTable)
-            table.clear()
-            topo_machines = topo.get("machines", {})
-            for mname in sorted(topo_machines):
-                minfo = topo_machines[mname]
-                addrs = minfo.get("addresses", {})
-                rconf = runtime_config.get(mname, {})
-                online = rconf.get("online")
-                if online:
-                    dot = "[green]\u25cf[/]"
-                elif rconf.get("ping"):
-                    dot = "[yellow]\u25cf[/]"
-                elif online is False:
-                    dot = "[red]\u25cf[/]"
-                else:
-                    dot = "[dim]\u25cb[/]"
-                ssh_conf = topo.get("ssh", {})
-                user = ssh_conf.get("base_station_user", "rrg") if minfo.get("role") == "base_station" else ssh_conf.get("robot_user", "swarm")
-                table.add_row(
-                    dot,
-                    mname,
-                    minfo.get("role", "?"),
-                    minfo.get("platform_id", "") or "—",
-                    addrs.get("mit_wifi", "—"),
-                    addrs.get("silvus", "—"),
-                    user,
-                    minfo.get("desc", ""),
-                    key=mname,
-                )
-
-        def _get_selected_name(self):
-            table = self.query_one("#config_table", DataTable)
-            try:
-                from textual.coordinate import Coordinate
-                row_idx = table.cursor_coordinate.row
-                return str(table.get_cell_at(Coordinate(row_idx, 1)))
-            except Exception:
-                return None
-
-        def _apply_result(self, result, original_name=None):
-            """Apply an EditMachineScreen result dict to topo and runtime_config."""
-            if not result:
-                return
-            log = self.query_one("#config_log", RichLog)
-            name = result["name"]
-            role = result["role"]
-            platform = result["platform"]
-            desc = result["desc"]
-            mit_wifi = result["mit_wifi"]
-            silvus = result["silvus"]
-            topo_machines = topo.setdefault("machines", {})
-            addrs = {}
-            if mit_wifi:
-                addrs["mit_wifi"] = mit_wifi
-            if silvus:
-                addrs["silvus"] = silvus
-            # Preserve other networks (e.g. penn_wifi) from the original entry
-            ref_name = original_name or name
-            if ref_name in topo_machines:
-                for net, ip in topo_machines[ref_name].get("addresses", {}).items():
-                    if net not in ("mit_wifi", "silvus"):
-                        addrs[net] = ip
-            entry = {"role": role, "addresses": addrs}
-            if platform:
-                entry["platform_id"] = platform
-            if desc:
-                entry["desc"] = desc
-            topo_machines[name] = entry
-            if active_network in addrs:
-                ssh_conf = topo.get("ssh", {})
-                user = ssh_conf.get("base_station_user", "rrg") if role == "base_station" else ssh_conf.get("robot_user", "swarm")
-                runtime_config[name] = {
-                    "name": name, "role": role, "ip": addrs[active_network],
-                    "user": user, "desc": desc, "platform_id": platform or "",
-                }
-            log.write(f"[green]Updated {name}[/]")
-            self._refresh_config_table()
-
-        def action_edit_selected(self):
-            name = self._get_selected_name()
-            if not name:
-                return
-            minfo = topo.get("machines", {}).get(name, {})
-            self.app.push_screen(EditMachineScreen(name, minfo), lambda r: self._apply_result(r, original_name=name))
-
-        def action_add_machine_action(self):
-            existing = set(topo.get("machines", {}).keys())
-            self.app.push_screen(EditMachineScreen(None, {}, existing_names=existing), self._apply_result)
-
-        def action_remove_selected(self):
-            log = self.query_one("#config_log", RichLog)
-            name = self._get_selected_name()
-            if not name:
-                log.write("[yellow]No machine selected.[/]")
-                return
-            topo_machines = topo.get("machines", {})
-            if name in topo_machines:
-                del topo_machines[name]
-            if name in runtime_config:
-                del runtime_config[name]
-            _deleted_machines.add(name)
-            log.write(f"Removed [bold]{name}[/]. Press F5 to save permanently.")
-            self._refresh_config_table()
-
-        def action_recheck_action(self):
-            log = self.query_one("#config_log", RichLog)
-            log.write("[dim]Rechecking connectivity...[/]")
-
-            def do_check():
-                machines = list(runtime_config.values())
-                reachable, unreachable = filter_reachable(machines)
-                return reachable + unreachable
-
-            def on_done(all_machines):
-                n_online = sum(1 for m in all_machines if m.get("online"))
-                log.write(f"[green]{n_online}/{len(all_machines)} machines online.[/]")
-                self._refresh_config_table()
-
-            def bg():
-                results = do_check()
-                self.app.call_from_thread(on_done, results)
-
-            threading.Thread(target=bg, daemon=True).start()
-
-        def action_save_action(self):
-            log = self.query_one("#config_log", RichLog)
-            topo_machines = topo.get("machines", {})
-            for mname in list(_deleted_machines):
-                if mname in topo_machines:
-                    del topo_machines[mname]
-            _deleted_machines.clear()
-            topo["machines"] = topo_machines
-            topo_path = pathlib.Path(topology_path) if topology_path else _DEFAULT_TOPOLOGY
-            try:
-                with open(topo_path, "w") as f:
-                    yaml.dump(topo, f, default_flow_style=False, sort_keys=False)
-                log.write(f"[green]Saved to {topo_path}[/]")
-            except OSError as e:
-                log.write(f"[red]Failed to save: {e}[/]")
-
-    # ---- Launch Screen ----
-
-    class LaunchScreen(ModalScreen):
-        BINDINGS = [
-            Binding("escape", "dismiss", "Back", priority=True),
-            Binding("m", "preset_mapping", "Mapping (m)", priority=True),
-            Binding("r", "preset_relocalize", "Relocalize (r)", priority=True),
-            Binding("d", "dry_run", "Dry Run", priority=True),
-            Binding("g", "do_launch", "Launch (g)", priority=True),
-            Binding("z", "deploy_zenoh", "Deploy Zenoh (z)", priority=True),
-            Binding("p", "browse_prior", "Refresh Priors (p)", priority=True),
-            Binding("f2", "toggle_log_size", "Log Size [F2]", priority=True),
-        ]
-
-        DEFAULT_CSS = """
-        LaunchScreen #robot_select {
-            max-height: 6;
-        }
-        LaunchScreen #session_select {
-            max-height: 12;
-        }
-        LaunchScreen #bs_select {
-            max-height: 4;
-        }
-        LaunchScreen #bs_session_select {
-            max-height: 6;
-        }
-        LaunchScreen #prior_tree {
-            min-height: 8;
-            max-height: 16;
-        }
-        LaunchScreen #launch_log {
-            min-height: 12;
-        }
-        """
-
-        _LOG_SIZES = [12, 24, 40]
-
-        def __init__(self, mode=None):
-            super().__init__()
-            self._launching = False
-            self._mode = mode  # None=manual, "mapping", "relocalize"
-            self._log_size_idx = 0
-
-        def compose(self) -> ComposeResult:
-            robots = [
-                (m["name"], m["name"], m.get("online", False))
-                for m in runtime_config.values()
-                if m["role"] == "robot"
-            ]
-            base_stations = [
-                (m["name"], f"{m['name']} (local)" if m["ip"] in _local_ips else m["name"], m.get("online", False) or m["ip"] in _local_ips)
-                for m in runtime_config.values()
-                if m["role"] == "base_station"
-            ]
-            self._all_sessions = sorted(
-                [f.stem for f in _TMUX_DIR.glob("*.yaml")]
-            ) if _TMUX_DIR.exists() else []
-
-            yield VerticalScroll(
-                Label("[bold]Launch[/] — m=mapping, r=relocalize, g=launch, d=dry run, p=refresh priors"),
-                Rule(),
-                Label("Select robots:"),
-                SelectionList(*robots, id="robot_select"),
-                Rule(),
-                Label("Robot session:"),
-                Input(placeholder="Filter sessions...", id="session_search"),
-                SelectionList(*[(s, s) for s in self._all_sessions], id="session_select"),
-                Rule(),
-                Label("Base station (optional):"),
-                SelectionList(*base_stations, id="bs_select"),
-                Label("Base station session:"),
-                SelectionList(
-                    *[(s, s) for s in self._all_sessions if "base_station" in s],
-                    id="bs_session_select",
-                ),
-                Rule(),
-                Label("Experiment (saved to ~/adt4_output/<name>/, e.g. 03252026_building_test):"),
-                Input(placeholder="date_description", id="experiment_input"),
-                Rule(),
-                PriorMapSelector(id="launch_prior_selector"),
-                Rule(),
-                Vertical(id="launch_progress_area"),
-                RichLog(id="launch_log", highlight=True, markup=True, max_lines=100),
-            )
-            yield Footer()
-
-        def on_mount(self):
-            if self._mode:
-                self._apply_preset(self._mode, update_robots=True)
-
-        def action_preset_mapping(self):
-            self._mode = "mapping"
-            # Save to app so re-opening preserves mode
-            try:
-                self.app._last_launch_mode = "mapping"
-            except Exception:
-                pass
-            self._apply_preset("mapping")
-
-        def action_preset_relocalize(self):
-            self._mode = "relocalize"
-            try:
-                self.app._last_launch_mode = "relocalize"
-            except Exception:
-                pass
-            self._apply_preset("relocalize")
-
-        def _apply_preset(self, mode, update_robots=False):
-            """Auto-populate selections from fleet_defaults in topology.
-
-            update_robots: if True, reset robot selection to all online robots
-            (only done on initial mount, not when the user switches mode).
-            """
-            defaults = topo.get("fleet_defaults", {})
-            spot_platforms = set(defaults.get("spot_platforms", ["smaug", "topaz"]))
-            sessions_cfg = defaults.get("sessions", {})
-            mode_sessions = sessions_cfg.get(mode, {})
-            spot_session = mode_sessions.get("spot", "")
-            phoenix_session = mode_sessions.get("phoenix", "")
-            bs_session = defaults.get("base_station_session", "")
-
-            log = self.query_one("#launch_log", RichLog)
-            log.write(f"[bold cyan]Auto-populating for {mode} mode...[/]")
-
-            # Select online robots (only on initial mount)
-            robot_sel = self.query_one("#robot_select", SelectionList)
-            if update_robots:
-                robot_sel.deselect_all()
-                for m in runtime_config.values():
-                    if m["role"] == "robot" and m.get("online"):
-                        robot_sel.select(m["name"])
-
-            # Pick session based on first selected (or online) robot's platform
-            selected_robots = list(robot_sel.selected)
-            first_robot = next(
-                (runtime_config[n] for n in selected_robots if n in runtime_config),
-                None,
-            ) or next(
-                (m for m in runtime_config.values()
-                 if m["role"] == "robot" and m.get("online")),
-                None,
-            )
-            target_session = ""
-            if first_robot:
-                target_session = spot_session if first_robot.get("platform_id") in spot_platforms else phoenix_session
-
-            if target_session:
-                sess_sel = self.query_one("#session_select", SelectionList)
-                sess_sel.deselect_all()
-                sess_sel.select(target_session)
-                log.write(f"  Robot session: [cyan]{target_session}[/]")
-                # Scroll to selected item
-                try:
-                    idx = self._all_sessions.index(target_session)
-                    sess_sel.scroll_to_center(sess_sel.get_option_at_index(idx))
-                except Exception:
-                    pass
-
-            # Select local base station
-            local_bs = None
-            bs_sel = self.query_one("#bs_select", SelectionList)
-            bs_sel.deselect_all()
-            for m in runtime_config.values():
-                if m["role"] == "base_station" and m["ip"] in _local_ips:
-                    bs_sel.select(m["name"])
-                    local_bs = m["name"]
-                    break
-            if local_bs:
-                log.write(f"  Base station: [cyan]{local_bs} (local)[/]")
-
-            # Select BS session
-            bs_sess_sel = self.query_one("#bs_session_select", SelectionList)
-            bs_sess_sel.deselect_all()
-            if bs_session:
-                bs_sess_sel.select(bs_session)
-                log.write(f"  BS session: [cyan]{bs_session}[/]")
-
-            if mode == "relocalize":
-                log.write("[yellow]Relocalize mode: prior map is required. Select one below or press p to refresh.[/]")
-            log.write("[dim]Review selections, then g=launch, d=dry run.[/]")
-
-        def on_input_changed(self, event: Input.Changed):
-            if event.input.id == "session_search":
-                query = event.value.strip()
-                sel = self.query_one("#session_select", SelectionList)
-                sel.clear_options()
-                for s in self._all_sessions:
-                    if _fuzzy_match(query, s):
-                        sel.add_option((s, s))
-
-        def _get_selected_session(self):
-            sel = self.query_one("#session_select", SelectionList)
-            selected = list(sel.selected)
-            return selected[0] if selected else ""
-
-        def action_browse_prior(self):
-            log = self.query_one("#launch_log", RichLog)
-            self.query_one("#launch_prior_selector", PriorMapSelector).refresh_priors(log)
-
-        def action_deploy_zenoh(self):
-            """Deploy zenoh config to base station (robots just listen)."""
-            log = self.query_one("#launch_log", RichLog)
-
-            # Get selected robots (for endpoint scoping)
-            robot_sel = self.query_one("#robot_select", SelectionList)
-            robot_names = list(robot_sel.selected)
-
-            # Only deploy to base station
-            bs_sel = self.query_one("#bs_select", SelectionList)
-            bs_names = list(bs_sel.selected)
-            if not bs_names or bs_names[0] not in runtime_config:
-                log.write("[yellow]No base station selected.[/]")
-                return
-
-            targets = [runtime_config[bs_names[0]]]
-            log.write(f"[bold]Deploying zenoh config to {targets[0]['name']}...[/]")
-            log.write(f"  [dim]Endpoints → robots: {', '.join(robot_names)}[/]")
-
-            def do_deploy():
-                results = []
-                for m in targets:
-                    try:
-                        endpoints = generate_zenoh_endpoints(
-                            topo, active_network, m["name"],
-                            robot_filter=robot_names,
-                        )
-                        ok, msg = deploy_zenoh_config(m["user"], m["ip"], endpoints)
-                        results.append((m["name"], ok, msg, len(endpoints)))
-                    except KeyError:
-                        results.append((m["name"], False, f"no address on {active_network}", 0))
-                return results
-
-            def on_done(results):
-                for name, ok, msg, n_ep in results:
-                    color = "green" if ok else "red"
-                    log.write(f"  [{color}]{name}: {msg} ({n_ep} endpoint(s))[/{color}]")
-                log.write("[bold]Zenoh deploy complete.[/]")
-
-            def bg():
-                results = do_deploy()
-                self.app.call_from_thread(on_done, results)
-
-            threading.Thread(target=bg, daemon=True).start()
-
-        def action_toggle_log_size(self):
-            self._log_size_idx = (self._log_size_idx + 1) % len(self._LOG_SIZES)
-            new_h = self._LOG_SIZES[self._log_size_idx]
-            log = self.query_one("#launch_log", RichLog)
-            log.styles.min_height = new_h
-
-        def _build_commands(self):
-            log = self.query_one("#launch_log", RichLog)
-            robot_sel = self.query_one("#robot_select", SelectionList)
-            selected = list(robot_sel.selected)
-            if not selected:
-                return [], None, "No robots selected."
-
-            session = self._get_selected_session()
-            if not session:
-                return [], None, "No session selected."
-
-            experiment = self.query_one("#experiment_input", Input).value.strip()
-            if not experiment:
-                return [], None, "No experiment name provided."
-
-            prior_path, prior_source = self.query_one("#launch_prior_selector", PriorMapSelector).get_prior_map()
-
-            if self._mode == "relocalize" and not prior_path:
-                return [], None, "Relocalize mode requires a prior map. Select one or press p to refresh."
-
-            # If prior map comes from a source machine, targets get it as prior_<name>
-            if prior_path and prior_source:
-                exp_name = pathlib.Path(prior_path).name
-                dest_prior = f"{output_root}/prior_{exp_name}"
-                # Source machine uses original path, others use transferred path
-                prior_paths = {}
-                for rname in selected:
-                    if rname == prior_source:
-                        prior_paths[rname] = prior_path
-                    else:
-                        prior_paths[rname] = dest_prior
-            else:
-                prior_paths = {rname: prior_path for rname in selected}
-
-            cmds = []
-            for rname in selected:
-                m = runtime_config[rname]
-                out_dir = f"{output_root}/{experiment}"
-                rprior = prior_paths.get(rname, "")
-                prior_flag = f" -p {_quote_path(rprior)}" if rprior else ""
-                remote_cmd = (
-                    f"cd ~/dcist_ws && source install/setup.zsh && "
-                    f"src/awesome_dcist_t4/dcist_launch_system/bin/run-adt4 "
-                    f"{shlex.quote(session)} -n {shlex.quote(m['name'])} "
-                    f"-c {shlex.quote(m['platform_id'])} "
-                    f"-o {_quote_path(out_dir)} -y{prior_flag} "
-                    f"--tmuxp-args \"-d\""
-                )
-                cmds.append((m, remote_cmd))
-
-            prior_info = (prior_path, prior_source) if prior_path else None
-            return cmds, prior_info, None
-
-
-        def action_dry_run(self):
-            log = self.query_one("#launch_log", RichLog)
-            cmds, prior_info, err = self._build_commands()
-            if err:
-                log.write(f"[yellow]{err}[/]")
-                return
-            log.write("[bold cyan]Dry run:[/]")
-            if prior_info:
-                prior_path, prior_source = prior_info
-                if prior_source:
-                    exp_name = pathlib.Path(prior_path).name
-                    dest_name = f"prior_{exp_name}"
-                    for m, cmd in cmds:
-                        if m["name"] != prior_source:
-                            log.write(
-                                f"  [dim]Transfer prior map: {prior_source} -> {m['name']} "
-                                f"({exp_name} as {dest_name})[/]"
-                            )
-                    log.write(f"  [dim]Prior map path on targets: {output_root}/{dest_name}[/]")
-            for m, cmd in cmds:
-                log.write(f"  ssh {m['user']}@{m['ip']} \"{cmd}\"")
-
-        def _transfer_prior_to_robot(self, prior_path, prior_source, target_machine, log, progress_callback=None):
-            """Transfer prior map from source machine to target if it doesn't exist."""
-            exp_name = pathlib.Path(prior_path).name
-            dest_name = f"prior_{exp_name}"
-            dest_path = f"{output_root}/{dest_name}"
-
-            # Check if prior already exists on target
-            rc, out, _ = ssh_cmd(
-                target_machine["user"], target_machine["ip"],
-                f"test -d {_quote_path(dest_path + '/hydra')} && test -d {_quote_path(dest_path + '/roman')} && echo OK",
-                timeout=10,
-            )
-            if "OK" in out:
-                self.app.call_from_thread(
-                    log.write,
-                    f"  [dim]{target_machine['name']}: prior map already present ({dest_name})[/]"
-                )
-                if progress_callback:
-                    progress_callback(100)
-                return True, dest_path
-
-            # Need to transfer
-            self.app.call_from_thread(
-                log.write,
-                f"  [cyan]Transferring prior map to {target_machine['name']}...[/]"
-            )
-            src = runtime_config.get(prior_source)
-            if not src:
-                self.app.call_from_thread(
-                    log.write,
-                    f"  [red]Source machine '{prior_source}' not found in config.[/]"
-                )
-                return False, ""
-
-            ok, method, msg = rsync_transfer(
-                src["user"], src["ip"], prior_path,
-                target_machine["user"], target_machine["ip"], dest_path,
-                relay=False, exclude=["recorded_data/"],
-                progress_callback=progress_callback,
-            )
-            if ok:
-                if progress_callback:
-                    progress_callback(100)
-                self.app.call_from_thread(
-                    log.write,
-                    f"  [green]{target_machine['name']}: prior map transferred ({method})[/]"
-                )
-                return True, dest_path
-            else:
-                self.app.call_from_thread(
-                    log.write,
-                    f"  [red]{target_machine['name']}: prior map transfer FAILED: {msg}[/]"
-                )
-                return False, ""
-
-        def _update_launch_progress(self, name, pct):
-            try:
-                bar = self.query_one(f"#lpb_{name}", ProgressBar)
-                bar.update(progress=pct)
-            except Exception:
-                pass
-
-        def action_do_launch(self):
-            if self._launching:
-                return
-            log = self.query_one("#launch_log", RichLog)
-            cmds, prior_info, err = self._build_commands()
-            if err:
-                log.write(f"[yellow]{err}[/]")
-                return
-
-            self._launching = True
-            experiment = self.query_one("#experiment_input", Input).value.strip()
-            # Snapshot session name before thread starts (DOM queries from threads are unsafe)
-            _launch_session_name = self._get_selected_session()
-
-            # Snapshot base station selection
-            robot_sel = self.query_one("#robot_select", SelectionList)
-            _robot_names = list(robot_sel.selected)
-
-            bs_sel = self.query_one("#bs_select", SelectionList)
-            bs_selected = list(bs_sel.selected)
-            bs_sess_sel = self.query_one("#bs_session_select", SelectionList)
-            bs_sess_selected = list(bs_sess_sel.selected)
-
-            # Debug: show raw selection values
-            log.write(f"[dim]robots.selected = {_robot_names}[/]")
-            log.write(f"[dim]session = {_launch_session_name}[/]")
-            log.write(f"[dim]bs.selected = {bs_selected}[/]")
-            log.write(f"[dim]bs_session.selected = {bs_sess_selected}[/]")
-            log.write(f"[dim]bs option_count = {bs_sel.option_count}[/]")
-
-            _bs_machine = runtime_config.get(bs_selected[0]) if bs_selected else None
-            _bs_session = bs_sess_selected[0] if bs_sess_selected else None
-
-            # Set up progress bars for transfers if needed
-            area = self.query_one("#launch_progress_area", Vertical)
-            area.remove_children()
-            if prior_info:
-                prior_path, prior_source = prior_info
-                if prior_source:
-                    targets = [m for m, _ in cmds if m["name"] != prior_source]
-                    if targets:
-                        area.mount(Label("[bold]Prior map transfer progress:[/]"))
-                        for m in targets:
-                            area.mount(
-                                Vertical(
-                                    Label(f"  {m['name']}:"),
-                                    ProgressBar(total=100, show_eta=False, id=f"lpb_{m['name']}"),
-                                )
-                            )
-
-            def do_launch():
-                try:
-                    # Phase 0: Deploy zenoh config to base station only
-                    # (robots just listen — star topology, no config change needed)
-                    deploy_targets = [_bs_machine] if _bs_machine else []
-                    if deploy_targets:
-                        self.app.call_from_thread(
-                            log.write, f"[bold]Deploying zenoh config to base station...[/]"
-                        )
-                        all_ok = True
-                        for m in deploy_targets:
-                            try:
-                                endpoints = generate_zenoh_endpoints(
-                                    topo, active_network, m["name"],
-                                    robot_filter=_robot_names,
-                                )
-                                ok, msg = deploy_zenoh_config(m["user"], m["ip"], endpoints)
-                                color = "green" if ok else "red"
-                                self.app.call_from_thread(
-                                    log.write, f"  [{color}]{m['name']}: {msg} ({len(endpoints)} ep)[/{color}]"
-                                )
-                                if not ok:
-                                    all_ok = False
-                            except KeyError:
-                                self.app.call_from_thread(
-                                    log.write, f"  [dim]{m['name']}: no address on {active_network}[/]"
-                                )
-                        if not all_ok:
-                            self.app.call_from_thread(
-                                log.write, "[bold red]ABORT: zenoh deploy failed on some machines.[/]"
-                            )
-                            return
-
-                    # Phase 1: Transfer prior maps if needed
-                    if prior_info:
-                        prior_path, prior_source = prior_info
-                        if prior_source:
-                            self.app.call_from_thread(
-                                log.write, "[bold]Phase 1: Ensuring prior maps on all robots...[/]"
-                            )
-                            for m, cmd in cmds:
-                                if m["name"] == prior_source:
-                                    continue  # source already has it
-
-                                def make_cb(mname):
-                                    def cb(pct):
-                                        self.app.call_from_thread(
-                                            self._update_launch_progress, mname, pct
-                                        )
-                                    return cb
-
-                                ok, dest_path = self._transfer_prior_to_robot(
-                                    prior_path, prior_source, m, log,
-                                    progress_callback=make_cb(m["name"]),
-                                )
-                                if not ok:
-                                    self.app.call_from_thread(
-                                        log.write,
-                                        f"  [bold red]ABORT: could not transfer prior to {m['name']}[/]"
-                                    )
-                                    return
-
-                    # Phase 1.5: Verify no conflicting tmux sessions
-                    all_targets = [m for m, _ in cmds]
-                    if _launch_session_name:
-                        if not _check_sessions_and_prompt(self.app, all_targets, _launch_session_name, log):
-                            return
-
-                    # Phase 2: Check for existing output dirs
-                    out_dir = f"{output_root}/{experiment}"
-                    conflicts = []
-                    for m, cmd in cmds:
-                        rc, out, _ = ssh_cmd(m["user"], m["ip"], f"test -d {_quote_path(out_dir)} && echo EXISTS", timeout=5)
-                        if "EXISTS" in out:
-                            conflicts.append(m["name"])
-
-                    if conflicts:
-                        names = ", ".join(conflicts)
-                        self.app.call_from_thread(
-                            log.write,
-                            f"[yellow]{out_dir} already exists on: {names}[/]"
-                        )
-                        # Prompt user via modal
-                        user_decision = [None]
-                        ev = threading.Event()
-
-                        def ask_overwrite():
-                            def on_dismiss(result):
-                                user_decision[0] = result
-                                ev.set()
-                            msg = (
-                                f"Output directory '{experiment}' already exists on:\n"
-                                f"{names}\n\n"
-                                f"Overwrite existing directories and launch?"
-                            )
-                            self.app.push_screen(
-                                ConfirmScreen(msg, yes_label="Yes, Overwrite (y)", no_label="No, Keep (n)"), on_dismiss
-                            )
-
-                        self.app.call_from_thread(ask_overwrite)
-                        ev.wait(timeout=120)
-
-                        if user_decision[0] is None:
-                            self.app.call_from_thread(
-                                log.write, "[red]Launch aborted — confirmation timed out or app closed.[/]"
-                            )
-                            return
-                        if not user_decision[0]:
-                            self.app.call_from_thread(
-                                log.write, "[red]Launch aborted — existing directories preserved.[/]"
-                            )
-                            return
-
-                        # Remove existing dirs on conflicting machines
-                        for m, cmd in cmds:
-                            if m["name"] in conflicts:
-                                self.app.call_from_thread(
-                                    log.write, f"  [dim]Removing {out_dir} on {m['name']}...[/]"
-                                )
-                                ssh_cmd(m["user"], m["ip"], f"rm -rf {_quote_path(out_dir)}", timeout=15)
-
-                    # Launch on each robot
-                    self.app.call_from_thread(
-                        log.write,
-                        "[bold]" + ("Phase 2: " if prior_info else "") + "Launching...[/]"
-                    )
-                    for m, cmd in cmds:
-                        self.app.call_from_thread(
-                            log.write, f"  Launching on [bold]{m['name']}[/]..."
-                        )
-                        rc, out, launch_err = ssh_cmd(m["user"], m["ip"], cmd, timeout=30)
-                        if rc == 0:
-                            self.app.call_from_thread(
-                                log.write, f"  [green]{m['name']}: launched[/]"
-                            )
-                        else:
-                            self.app.call_from_thread(
-                                log.write, f"  [red]{m['name']}: FAILED (rc={rc}) {launch_err}[/]"
-                            )
-                    # Phase 3: Launch base station if selected
-                    if _bs_machine and _bs_session:
-                        self.app.call_from_thread(
-                            log.write, "[bold]Launching base station...[/]"
-                        )
-                        # Stop fleet zenohd so BS can bind port 7447
-                        if _fleet_zenoh["proc"]:
-                            self.app.call_from_thread(
-                                log.write, "  [dim]Stopping fleet zenohd (base station will take over)...[/]"
-                            )
-                            _stop_fleet_zenohd()
-
-                        # Generate namespaced rviz config:
-                        # mapping → first selected robot; relocalize → base station
-                        # (BS publishes the prior map in relocalize mode)
-                        rviz_flag = ""
-                        try:
-                            if self._mode == "relocalize":
-                                rviz_name = _bs_machine["name"]
-                            elif _robot_names:
-                                rviz_name = _robot_names[0]
-                            else:
-                                rviz_name = None
-                            if rviz_name:
-                                rviz_path = generate_namespaced_rviz(rviz_name)
-                                rviz_flag = f" --rviz-config {shlex.quote(rviz_path)}"
-                                self.app.call_from_thread(
-                                    log.write, f"  [dim]RViz config: {rviz_path}[/]"
-                                )
-                        except Exception as e:
-                            self.app.call_from_thread(
-                                log.write, f"  [yellow]RViz config generation failed: {e}[/]"
-                            )
-
-                        out_dir = f"{output_root}/{experiment}"
-                        q_output = _quote_path(out_dir)
-                        platform = _bs_machine.get("platform_id") or _bs_machine["name"]
-                        config_flag = f" -c {shlex.quote(platform)}" if _bs_machine.get("platform_id") else ""
-                        bs_cmd = (
-                            f"cd ~/dcist_ws && source install/setup.zsh && "
-                            f"nohup src/awesome_dcist_t4/dcist_launch_system/bin/run-adt4 "
-                            f"{shlex.quote(_bs_session)} -n {shlex.quote(_bs_machine['name'])}"
-                            f"{config_flag} -o {q_output} -y{rviz_flag} "
-                            f"--tmuxp-args \"-d\" "
-                            f">/dev/null 2>&1 &"
-                        )
-                        self.app.call_from_thread(
-                            log.write, f"  [dim][{_bs_machine['user']}@{_bs_machine['ip']}] {bs_cmd}[/]"
-                        )
-                        rc, out, err = ssh_cmd(_bs_machine["user"], _bs_machine["ip"], bs_cmd, timeout=60)
-                        # nohup & always returns 0; verify tmux session started
-                        import time
-                        time.sleep(3)
-                        tmux_name = "adt4_system"
-                        trc, tout, _ = ssh_cmd(
-                            _bs_machine["user"], _bs_machine["ip"],
-                            f"tmux has-session -t {shlex.quote(tmux_name)} 2>/dev/null && echo OK",
-                            timeout=5,
-                        )
-                        if "OK" in tout:
-                            self.app.call_from_thread(
-                                log.write, f"  [green]{_bs_machine['name']}: base station launched (tmux session: {tmux_name})[/]"
-                            )
-                            self.app.call_from_thread(
-                                log.write, "  [dim]Base station zenohd is now active on port 7447.[/]"
-                            )
-                        else:
-                            self.app.call_from_thread(
-                                log.write, f"  [red]{_bs_machine['name']}: tmux session '{tmux_name}' not found after launch. rc={rc} {err}[/]"
-                            )
-
-                    self.app.call_from_thread(log.write, "[bold]Launch sequence complete.[/]")
-                finally:
-                    self._launching = False
-
-            threading.Thread(target=do_launch, daemon=True).start()
-
-    # ---- Bandwidth Selection Screen ----
-
-    class BandwidthSelectScreen(ModalScreen):
-        """Select robots, run iperf3 tests with live progress, dismiss with results."""
-        AUTO_FOCUS = "SelectionList"
-        BINDINGS = [
-            Binding("escape", "close", "Cancel/Close", priority=True),
-            Binding("f5", "run_test", "Run [F5]", priority=True),
-        ]
-        DEFAULT_CSS = """
-        BandwidthSelectScreen { align: center middle; }
-        BandwidthSelectScreen > Vertical {
-            width: 60;
-            height: auto;
-            max-height: 35;
-            border: thick $primary;
-            background: $surface;
-            padding: 1 2;
-        }
-        BandwidthSelectScreen #bw_progress { height: 15; }
-        """
-
-        def __init__(self, machines, server_ip, **kwargs):
-            super().__init__(**kwargs)
-            self._machines = machines
-            self._server_ip = server_ip
-            self._results = []
-            self._done = False
-
-        def compose(self) -> ComposeResult:
-            options = [(m["name"], m["name"], True) for m in self._machines]
-            yield Vertical(
-                Label("[bold]Bandwidth Test[/]"),
-                Label("[dim]Space=toggle  F5=run  Escape=cancel[/]", id="bw_hint"),
-                Rule(),
-                SelectionList(*options, id="bw_select"),
-                RichLog(id="bw_progress", highlight=True, markup=True, max_lines=100),
-            )
-
-        def on_mount(self):
-            self.query_one("#bw_progress", RichLog).display = False
-
-        def action_close(self):
-            self.dismiss(self._results or None)
-
-        def action_run_test(self):
-            if self._done:
-                return
-            sel = self.query_one("#bw_select", SelectionList)
-            chosen = [m for m in self._machines if m["name"] in set(sel.selected)]
-            if not chosen:
-                return
-
-            # Switch to progress view
-            sel.display = False
-            hint = self.query_one("#bw_hint", Label)
-            hint.update("[dim]Escape to close when finished[/]")
-            plog = self.query_one("#bw_progress", RichLog)
-            plog.display = True
-            plog.write(f"[dim]Server: {self._server_ip}  →  {', '.join(m['name'] for m in chosen)}[/]")
-
-            def emit(r):
-                name = r.get("name", "?")
-                if "error" in r:
-                    plog.write(f"[red]{name}: {r['error']}[/]")
-                else:
-                    mbps = r.get("mbps", 0)
-                    color = "green" if mbps > 10 else "yellow" if mbps > 1 else "red"
-                    tunnel_warn = "  [yellow](via SSH tunnel — direct port blocked)[/]" if r.get("tunneled") else ""
-                    plog.write(
-                        f"[{color}]{name}: {mbps:.1f} Mbps[/]"
-                        f"  retransmits={r.get('retransmits', 0)}"
-                        f"  ({r.get('duration', 0):.1f}s)"
-                        f"{tunnel_warn}"
-                    )
-                self._results.append(r)
-
-            def all_done():
-                plog.write("[dim]Done — press Escape to close[/]")
-                self._done = True
-
-            def bg():
-                for m in chosen:
-                    name = m["name"]
-                    self.app.call_from_thread(plog.write, f"[bold]{name}[/]")
-                    if not check_iperf3(m["user"], m["ip"]):
-                        r = {"name": name, "error": "iperf3 not installed"}
-                    else:
-                        def on_interval(t_end, mbps, retransmits, _name=name):
-                            color = "green" if mbps > 10 else "yellow" if mbps > 1 else "red"
-                            retr = f"  retr={retransmits}" if retransmits else ""
-                            self.app.call_from_thread(
-                                plog.write,
-                                f"  [{color}]{t_end:.0f}s  {mbps:.0f} Mbits/sec{retr}[/]",
-                            )
-                        r = run_iperf3_test(
-                            self._server_ip, m["user"], m["ip"],
-                            interval_cb=on_interval,
-                        )
-                        r["name"] = name
-                    self.app.call_from_thread(emit, r)
-                self.app.call_from_thread(all_done)
-
-            threading.Thread(target=bg, daemon=True).start()
-
-    # ---- Monitor Screen ----
-
-    class MonitorScreen(ModalScreen):
-        BINDINGS = [
-            Binding("escape", "dismiss", "Back", priority=True),
-            Binding("r", "refresh", "Refresh", priority=True),
-            Binding("n", "refresh_nodes", "Node Status", priority=True),
-            Binding("k", "stop_save", "Stop & Save", priority=True),
-            Binding("enter", "drill_down", "Details", priority=True),
-            Binding("v", "gen_rviz", "RViz Config", priority=True),
-            Binding("z", "check_zenoh", "Zenoh/Radio", priority=True),
-            Binding("f3", "check_silvus_routes", "Check Routes [F3]", priority=True),
-            Binding("f4", "silvus_detail", "Radio Detail [F4]", priority=True),
-            Binding("f2", "start_zenohd", "Start Zenohd [F2]", priority=True),
-            Binding("w", "bandwidth_test", "Bandwidth [w]", priority=True),
-        ]
-
-        def compose(self) -> ComposeResult:
-            yield VerticalScroll(
-                Label("[bold]Fleet Monitor[/]"),
-                Rule(),
-                Label("[bold]System Status[/] (r=refresh)"),
-                Static(id="monitor_table"),
-                Rule(),
-                Label("[bold]Node Readiness[/] (n=refresh, Enter=details, Esc=back, k=stop & save)"),
-                DataTable(id="node_summary_table"),
-                Static(id="node_detail_view"),
-                Rule(),
-                RichLog(id="monitor_log", highlight=True, markup=True, max_lines=30),
-            )
-            yield Footer()
-
-        def _build_status_table(self, results=None):
-            """Build a Rich Table from system status results."""
-            t = Table(box=None, expand=True, pad_edge=False)
-            t.add_column("Machine", no_wrap=True, min_width=8)
-            t.add_column("Tmux", no_wrap=False, max_width=20)
-            t.add_column("ROS2", no_wrap=True, max_width=5)
-            t.add_column("Load 1/5/15m", no_wrap=False, max_width=16)
-            t.add_column("Mem", no_wrap=True, max_width=12)
-            t.add_column("GPU", no_wrap=False, max_width=18)
-            t.add_column("Disk", no_wrap=False, max_width=20)
-            t.add_column("Bat", no_wrap=True, max_width=18)
-            t.add_column("Radio", no_wrap=True, max_width=14)
-            if results:
-                for m, st in sorted(results, key=lambda x: x[0]["name"]):
-                    if isinstance(st, Exception):
-                        t.add_row(m["name"], str(st), "", "", "", "", "", "", "")
-                    else:
-                        t.add_row(
-                            m["name"],
-                            st["tmux_sessions"],
-                            st["ros2_procs"],
-                            st["load"],
-                            st.get("mem", "N/A"),
-                            st.get("gpu", "N/A"),
-                            st["disk"],
-                            st.get("battery", "N/A"),
-                            st.get("radio", "N/A"),
-                        )
-            return t
-
-        def on_mount(self):
-            self._node_status = {}  # robot_name -> [node dicts]
-
-            # Render empty table initially
-            widget = self.query_one("#monitor_table", Static)
-            widget.update(self._build_status_table())
-
-            node_table = self.query_one("#node_summary_table", DataTable)
-            node_table.add_column("Robot", key="robot")
-            node_table.add_column("Readiness", key="readiness")
-            node_table.add_column("Nodes", key="nodes")
-            node_table.add_column("Issues", key="issues")
-            node_table.cursor_type = "row"
-
-            self._node_poller = NodeStatusPoller()
-            self._node_poller.start()
-            self.action_refresh()
-            # Read cached poller data every 1s (instant — just reads memory)
-            self.set_interval(1, self._update_node_table)
-
-        def action_refresh(self):
-            log = self.query_one("#monitor_log", RichLog)
-            log.write("[dim]Refreshing system status...[/]")
-
-            def do_fetch():
-                machines = [m for m in runtime_config.values() if m.get("online", False)]
-                sys_radios = topo.get("silvus_radios", {}).get("radios", {})
-                silvus_ips = [r["mgmt_ip"] for r in sys_radios.values() if "mgmt_ip" in r]
-
-                def fetch(m):
-                    return get_remote_status(m["user"], m["ip"], silvus_ips=silvus_ips)
-
-                return run_parallel(fetch, machines)
-
-            def on_done(results):
-                widget = self.query_one("#monitor_table", Static)
-                widget.update(self._build_status_table(results))
-                log.write("[green]System status updated.[/]")
-
-            def bg():
-                results = do_fetch()
-                self.app.call_from_thread(on_done, results)
-
-            threading.Thread(target=bg, daemon=True).start()
-
-        def action_refresh_nodes(self):
-            """Manual refresh — restart poller if dead, show status."""
-            log = self.query_one("#monitor_log", RichLog)
-            if not self._node_poller.is_running():
-                ok, msg = self._node_poller.start()
-                log.write(f"[dim]Node poller: {msg}[/]")
-            err = self._node_poller.get_error()
-            if err:
-                log.write(f"[dim]ros2 stderr: {err}[/]")
-            self._update_node_table()
-            status = self._node_status
-            if status:
-                log.write(f"[green]Node status: {len(status)} machine(s)[/]")
-            else:
-                log.write("[yellow]No node status yet — poller may still be starting.[/]")
-
-        def _update_node_table(self):
-            """Read latest from persistent poller and refresh the table.
-
-            Uses update_cell to avoid clearing/rebuilding which resets cursor.
-            """
-            status = self._node_poller.get_status()
-            self._node_status = status
-            node_table = self.query_one("#node_summary_table", DataTable)
-
-            if not status:
-                return
-
-            # Build new row data
-            new_rows = {}
-            for robot_name in sorted(status):
-                nodes = status[robot_name]
-                color, label = compute_robot_readiness(nodes)
-                n_total = len(nodes)
-                n_nominal = sum(1 for n in nodes if n["status"] == _STATUS_NOMINAL)
-                issues = []
-                for n in nodes:
-                    if n["status"] != _STATUS_NOMINAL:
-                        sname = _STATUS_NAMES.get(n["status"], "?")
-                        issues.append(f"{n['nickname']}={sname}")
-                new_rows[robot_name] = (
-                    f"[{color}]{label}[/{color}]",
-                    f"{n_nominal}/{n_total}",
-                    ", ".join(issues) if issues else "none",
-                )
-
-            # Check if rows changed — update in place if possible
-            existing_keys = set()
-            try:
-                from textual.coordinate import Coordinate
-                for i in range(node_table.row_count):
-                    existing_keys.add(str(node_table.get_cell_at(Coordinate(i, 0))))
-            except Exception:
-                existing_keys = set()
-
-            if existing_keys == set(new_rows.keys()):
-                # Same robots — update cells in place (no cursor reset)
-                for i, robot_name in enumerate(sorted(new_rows)):
-                    readiness, node_count, issue_str = new_rows[robot_name]
-                    try:
-                        from textual.coordinate import Coordinate
-                        node_table.update_cell_at(Coordinate(i, 1), readiness)
-                        node_table.update_cell_at(Coordinate(i, 2), node_count)
-                        node_table.update_cell_at(Coordinate(i, 3), issue_str)
-                    except Exception:
-                        break
-            else:
-                # Robot set changed — rebuild
-                self._rebuilding_table = True
-                node_table.clear()
-                for robot_name in sorted(new_rows):
-                    readiness, node_count, issue_str = new_rows[robot_name]
-                    node_table.add_row(robot_name, readiness, node_count, issue_str, key=robot_name)
-                self._rebuilding_table = False
-
-            # Also refresh the detail view if one is showing
-            self._show_node_detail()
-
-        def on_data_table_cursor_moved(self, event):
-            """Update detail view when cursor moves via arrow keys or click."""
-            if event.data_table.id == "node_summary_table" and not getattr(self, "_rebuilding_table", False):
-                self._show_node_detail()
-
-        def action_drill_down(self):
-            self._show_node_detail()
-
-        def _show_node_detail(self):
-            """Show detailed per-node status for the currently selected robot."""
-            node_table = self.query_one("#node_summary_table", DataTable)
-            detail = self.query_one("#node_detail_view", Static)
-
-            try:
-                from textual.coordinate import Coordinate
-                row_idx = node_table.cursor_coordinate.row
-                robot_name = str(node_table.get_cell_at(Coordinate(row_idx, 0)))
-            except Exception:
-                return
-
-            nodes = self._node_status.get(robot_name)
-            if not nodes:
-                detail.update(f"[dim]No node data for {robot_name}.[/]")
-                return
-
-            t = Table(box=None, expand=True, pad_edge=False, title=f"Node details: {robot_name}")
-            t.add_column("Status", no_wrap=True, min_width=8)
-            t.add_column("Node", no_wrap=True, min_width=12)
-            t.add_column("Notes", no_wrap=False)
-            for n in sorted(nodes, key=lambda x: x["nickname"]):
-                status = n["status"]
-                color = _STATUS_COLORS.get(status, "white")
-                sname = _STATUS_NAMES.get(status, "?")
-                notes = n.get("notes", "")
-                t.add_row(f"[{color}]{sname}[/{color}]", n["nickname"], notes)
-            detail.update(t)
-
-        def action_stop_save(self):
-            """Send shutdown signal to all online robots via tmux."""
-            log = self.query_one("#monitor_log", RichLog)
-            online_machines = [m for m in runtime_config.values() if m.get("online", False)]
-            if not online_machines:
-                log.write("[red]No online machines to stop.[/]")
-                return
-
-            names = ", ".join(m["name"] for m in sorted(online_machines, key=lambda x: x["name"]))
-            log.write(f"[yellow]Sending Stop & Save to: {names}[/]")
-
-            def do_stop():
-                def stop_one(m):
-                    ok, msg = send_tmux_keys(m["user"], m["ip"])
-                    return m["name"], ok, msg
-                return run_parallel(stop_one, online_machines)
-
-            def on_done(results):
-                for item, result in sorted(results, key=lambda x: x[0]["name"]):
-                    if isinstance(result, Exception):
-                        log.write(f"  [red]{item['name']}: {result}[/]")
-                    else:
-                        mname, ok, msg = result
-                        color = "green" if ok else "red"
-                        log.write(f"  [{color}]{mname}: {msg}[/{color}]")
-                log.write("[bold]Stop & Save complete.[/]")
-
-            def bg():
-                results = do_stop()
-                self.app.call_from_thread(on_done, results)
-
-            threading.Thread(target=bg, daemon=True).start()
-
-        def action_gen_rviz(self):
-            """Generate a namespaced rviz config for a robot."""
-            log = self.query_one("#monitor_log", RichLog)
-            # Use the first robot in runtime_config as default
-            robots = [m for m in sorted(runtime_config) if runtime_config[m].get("role") == "robot"]
-            if not robots:
-                log.write("[red]No robots in config.[/]")
-                return
-
-            # Generate for each robot
-            for robot_name in robots:
-                mconf = runtime_config[robot_name]
-                platform_id = mconf.get("platform_id", robot_name)
-                out_path = generate_namespaced_rviz(robot_name)
-                log.write(f"  Generated [cyan]{out_path}[/] for {robot_name}")
-
-            log.write(f"[bold]RViz configs generated.[/] Launch with:")
-            log.write(f"  [dim]rviz2 -d /tmp/dcist_<robot>.rviz[/]")
-
-        def action_check_zenoh(self):
-            """Check zenoh port reachability + Silvus radio link quality for all machines."""
-            log = self.query_one("#monitor_log", RichLog)
-            log.write("[dim]Checking zenoh connectivity and radio link quality...[/]")
-
-            zenoh_conf = topo.get("zenoh", {})
-            zenoh_port = zenoh_conf.get("port", 7447)
-
-            zenoh_admin_port = zenoh_conf.get("admin_port", 8000)
-
-            def do_check():
-                machines = list(runtime_config.values())
-                sys_radios = topo.get("silvus_radios", {}).get("radios", {})
-                silvus_mgmt_ips = [r["mgmt_ip"] for r in sys_radios.values() if "mgmt_ip" in r]
-
-                results = []
-                for m in machines:
-                    ip = m["ip"]
-                    zenoh_ok = check_zenoh_port(ip, port=zenoh_port)
-
-                    radio = None
-                    if m.get("online") and silvus_mgmt_ips:
-                        try:
-                            radio = get_silvus_link_quality(m["user"], ip, silvus_mgmt_ips)
-                        except Exception:
-                            pass
-
-                    results.append((m, zenoh_ok, radio))
-
-                # Query Zenoh admin API from background thread
-                import urllib.request as _urlreq
-                zenoh_peers = None
-                zenoh_peers_err = None
-                try:
-                    url = f"http://localhost:{zenoh_admin_port}/@/router/local/link/**"
-                    with _urlreq.urlopen(url, timeout=2) as resp:
-                        zenoh_peers = json.loads(resp.read())
-                except Exception as e:
-                    zenoh_peers_err = str(e)
-
-                return results, zenoh_peers, zenoh_peers_err
-
-            def on_done(results, zenoh_peers, zenoh_peers_err):
-                log.write(f"\n[bold]Zenoh Port {zenoh_port} Reachability:[/]")
-                for m, zenoh_ok, _ in sorted(results, key=lambda x: x[0]["name"]):
-                    color = "green" if zenoh_ok else "red"
-                    status = "OPEN" if zenoh_ok else "CLOSED"
-                    log.write(f"  [{color}]{m['name']:12s} {m['ip']:18s} {status}[/{color}]")
-
-                has_radio = any(r for _, _, r in results if r)
-                if has_radio:
-                    log.write(f"\n[bold]Silvus Radio Link Quality:[/]")
-                    radio_assignment = []  # (machine_name, node_id)
-                    for m, _, radio in sorted(results, key=lambda x: x[0]["name"]):
-                        if not radio:
-                            continue
-                        if "error" in radio:
-                            log.write(f"  [red]{m['name']:12s} {radio['error']}[/]")
-                        elif "raw" in radio:
-                            log.write(f"  [dim]{m['name']:12s} {radio['raw']}[/]")
-                        else:
-                            bat = radio.get("battery", "?")
-                            mesh = radio.get("mesh_nodes", "?")
-                            neighbors = radio.get("neighbor_count", "?")
-                            rssi = radio.get("rssi", "N/A")
-                            snr = radio.get("snr", "N/A")
-                            log.write(
-                                f"  [cyan]{m['name']:12s}[/] "
-                                f"Bat:{bat}% Mesh:{mesh} Neighbors:{neighbors} "
-                                f"RSSI:{rssi} SNR:{snr}"
-                            )
-                            nid = radio.get("node_id")
-                            if nid:
-                                radio_assignment.append((m["name"], nid))
-
-                    if radio_assignment:
-                        log.write(f"\n[bold]Radio Assignment (current):[/]")
-                        for mname, nid in radio_assignment:
-                            log.write(f"  [cyan]{mname:12s}[/] → node {nid}")
-                        log.write(f"  [dim]Press F4 to see RSSI matrix and mesh topology.[/]")
-
-                # ---- Zenoh peer browser ----
-                log.write(f"\n[bold]Zenoh Connected Peers:[/]")
-                if zenoh_peers is not None:
-                    if zenoh_peers:
-                        for zid, meta in zenoh_peers.items():
-                            whatami = meta.get("whatami", "?") if isinstance(meta, dict) else "?"
-                            short_zid = zid[:8] if len(zid) > 8 else zid
-                            log.write(f"  [green]{short_zid}[/] ({whatami})")
-                    else:
-                        log.write("  [dim]No peers connected.[/]")
-                else:
-                    err_str = zenoh_peers_err or ""
-                    if "refused" in err_str or "timed out" in err_str or "unreachable" in err_str:
-                        log.write(f"  [dim]Admin API not available — start zenohd with --rest-http-port {zenoh_admin_port}[/]")
-                    else:
-                        log.write(f"  [dim]Admin API error: {err_str[:60]}[/]")
-
-                log.write("[bold]Zenoh/radio check complete.[/]")
-
-            def bg():
-                results, zenoh_peers, zenoh_peers_err = do_check()
-                self.app.call_from_thread(on_done, results, zenoh_peers, zenoh_peers_err)
-
-            threading.Thread(target=bg, daemon=True).start()
-
-        def action_check_silvus_routes(self):
-            """Check if Silvus management route exists on all machines + locally."""
-            log = self.query_one("#monitor_log", RichLog)
-            log.write("[dim]Checking Silvus management routes...[/]")
-
-            mgmt_subnet = topo.get("silvus_radios", {}).get("mgmt_subnet", "172.20.0.0/16")
-
-            def do_check():
-                results = []
-                # Check local machine first
-                local_info = check_silvus_route(mgmt_subnet=mgmt_subnet)
-                results.append(("LOCAL", None, local_info))
-
-                # Check all online machines
-                machines = [m for m in runtime_config.values() if m.get("online")]
-                for m in machines:
-                    try:
-                        info = check_silvus_route(m["user"], m["ip"], mgmt_subnet)
-                        results.append((m["name"], m, info))
-                    except Exception as e:
-                        results.append((m["name"], m, {"has_route": False, "interface": None, "silvus_iface": None, "error": str(e)}))
-                return results
-
-            def on_done(results):
-                log.write(f"\n[bold]Silvus Management Route ({mgmt_subnet}):[/]")
-                missing = []
-                for name, m, info in sorted(results, key=lambda x: x[0]):
-                    iface = info.get("silvus_iface") or "—"
-                    if info.get("has_route"):
-                        route = info.get("route_info", "OK")
-                        log.write(f"  [green]{name:12s}[/] iface={iface}  [green]ROUTE OK[/] ({route})")
-                    elif info.get("silvus_iface"):
-                        log.write(f"  [red]{name:12s}[/] iface={iface}  [red]NO ROUTE[/]")
-                        missing.append((name, m, iface))
-                    else:
-                        log.write(f"  [dim]{name:12s}[/] [dim]No Silvus interface (no 192.168.100.x address)[/]")
-                if missing:
-                    log.write(f"\n[bold yellow]Run these commands to add missing routes:[/]")
-                    for name, m, iface in missing:
-                        if m is None:
-                            # local machine
-                            log.write(f"\n  [bold]{name}[/] (run locally):")
-                            log.write(f"    [cyan]sudo ip route add {mgmt_subnet} dev {iface}[/]")
-                            log.write(f"  [dim]To make permanent (same USB dongle):[/]")
-                            log.write(f"    [cyan]conn=$(nmcli -f NAME,DEVICE con show | grep {iface} | awk '{{print $1}}')[/]")
-                            log.write(f"    [cyan]nmcli con modify \"$conn\" +ipv4.routes \"{mgmt_subnet}\"[/]")
-                            log.write(f"    [cyan]nmcli con up \"$conn\"[/]")
-                        else:
-                            user, ip = m["user"], m["ip"]
-                            log.write(f"\n  [bold]{name}[/] (SSH to {ip}):")
-                            log.write(f"    [cyan]ssh {user}@{ip} sudo ip route add {mgmt_subnet} dev {iface}[/]")
-                            log.write(f"  [dim]To make permanent (same USB dongle):[/]")
-                            log.write(f"    [cyan]ssh {user}@{ip}[/]")
-                            log.write(f"    [cyan]  conn=$(nmcli -f NAME,DEVICE con show | grep {iface} | awk '{{print $1}}')[/]")
-                            log.write(f"    [cyan]  nmcli con modify \"$conn\" +ipv4.routes \"{mgmt_subnet}\"[/]")
-                            log.write(f"    [cyan]  nmcli con up \"$conn\"[/]")
-
-            def bg():
-                results = do_check()
-                self.app.call_from_thread(on_done, results)
-
-            threading.Thread(target=bg, daemon=True).start()
-
-        def action_silvus_detail(self):
-            """Show Silvus RSSI link quality matrix and mesh routing topology."""
-            log = self.query_one("#monitor_log", RichLog)
-            sys_radios = topo.get("silvus_radios", {}).get("radios", {})
-            if not sys_radios:
-                log.write("[yellow]No Silvus radios defined in topology.[/]")
-                return
-            log.write("[dim]Querying Silvus radios (direct HTTP)...[/]")
-
-            def do_query():
-                return get_silvus_radio_details(topo)
-
-            def on_done(details):
-                if not details:
-                    log.write("[red]No radio data returned.[/]")
-                    return
-
-                # Build node_id → radio_name reverse map
-                nid_to_name = {}
-                for rname, info in details.items():
-                    nid = info.get("node_id")
-                    if nid:
-                        nid_to_name[int(nid)] = rname
-
-                names = sorted(details.keys())
-                col_w = max(len(n) for n in names) + 2
-
-                # ---- RSSI Matrix ----
-                log.write(f"\n[bold]Silvus RSSI Link Quality:[/]")
-                header = " " * (col_w + 2) + "".join(f"{n:>{col_w}}" for n in names)
-                log.write(f"  [dim]{header}[/]")
-
-                for src in names:
-                    info = details[src]
-                    if info.get("error"):
-                        log.write(f"  [red]{src:<{col_w}}[/]  [red](unreachable: {info['error'][:40]})[/]")
-                        continue
-                    row = f"  [cyan]{src:<{col_w}}[/]"
-                    link_stats = info.get("link_stats")
-                    rssi_raw = info.get("rssi")
-                    routing_tree = [int(x) for x in (info.get("routing_tree") or [])]
-                    src_nid = info.get("node_id")
-
-                    for dst in names:
-                        if src == dst:
-                            row += f"{'---':>{col_w}}"
-                            continue
-
-                        dst_nid = details[dst].get("node_id")
-                        val = None
-
-                        # Try link_stats (keyed by node_id string)
-                        if isinstance(link_stats, dict) and dst_nid is not None:
-                            try:
-                                val = float(link_stats.get(str(dst_nid)) or link_stats.get(str(int(dst_nid))))
-                            except (TypeError, ValueError):
-                                pass
-
-                        # Fallback: rssi field (single-neighbor case)
-                        if val is None and rssi_raw and isinstance(rssi_raw, list) and len(names) == 2:
-                            try:
-                                val = float(rssi_raw[0]) / 2.0
-                            except (TypeError, ValueError):
-                                pass
-
-                        if val is not None:
-                            if val > -60:
-                                cell = f"[green]{val:.0f}dB[/]"
-                            elif val > -80:
-                                cell = f"[yellow]{val:.0f}dB[/]"
-                            else:
-                                cell = f"[red]{val:.0f}dB[/]"
-                            row += f"{cell:>{col_w}}"
-                        elif dst_nid is not None and int(dst_nid) in routing_tree:
-                            row += f"[green]{'linked':>{col_w}}[/]"
-                        else:
-                            row += f"[dim]{'?':>{col_w}}[/]"
-                    log.write(row)
-
-                # ---- Mesh Topology ----
-                log.write(f"\n[bold]Mesh Routing (per radio):[/]")
-                for rname in names:
-                    info = details[rname]
-                    if info.get("error"):
-                        continue
-                    nid = info.get("node_id", "?")
-                    tree = info.get("routing_tree") or []
-                    # Resolve node IDs to radio names where possible
-                    tree_str = ", ".join(
-                        nid_to_name.get(int(n), str(n)) for n in tree
-                    ) if tree else "—"
-                    bat = info.get("battery")
-                    bat_str = f"  bat={bat:.0f}%" if bat is not None else ""
-                    log.write(f"  [cyan]{rname}[/] (node {nid}){bat_str}: [{tree_str}]")
-
-                mgmt_subnet = topo.get("silvus_radios", {}).get("mgmt_subnet", "172.20.0.0/16")
-                all_err = all(d.get("error") for d in details.values())
-                if all_err:
-                    log.write(f"\n[yellow]All radios unreachable — is the {mgmt_subnet} route set? (press F3)[/]")
-                else:
-                    log.write(f"\n[dim]Press 'z' to see which robot currently has each radio.[/]")
-
-            def bg():
-                result = do_query()
-                self.app.call_from_thread(on_done, result)
-
-            threading.Thread(target=bg, daemon=True).start()
-
-        def action_start_zenohd(self):
-            """Manually start/restart the fleet-managed zenohd."""
-            log = self.query_one("#monitor_log", RichLog)
-
-            # Check if a system zenohd is already running (e.g. from base station tmux)
-            system_zenohd = subprocess.run(
-                ["pgrep", "-f", "rmw_zenohd"], capture_output=True
-            ).returncode == 0
-            fleet_zenohd = _fleet_zenoh["proc"] and _fleet_zenoh["proc"].poll() is None
-
-            if system_zenohd and not fleet_zenohd:
-                log.write("[green]Base station zenohd is running on port 7447 — using it for ROS monitoring.[/]")
-                log.write("[dim]Press 'n' to refresh node status.[/]")
-                return
-            elif fleet_zenohd:
-                log.write("[dim]Restarting fleet zenohd...[/]")
-                _stop_fleet_zenohd()
-
-            ok, msg = _start_fleet_zenohd()
-            color = "green" if ok else "red"
-            log.write(f"[{color}]Fleet zenohd: {msg}[/{color}]")
-            if ok:
-                log.write("[dim]Press 'n' to refresh node status.[/]")
-
-        def action_bandwidth_test(self):
-            """Run iperf3 throughput test from base station to selected online robots."""
-            import shutil
-            log = self.query_one("#monitor_log", RichLog)
-
-            machines = [m for m in runtime_config.values()
-                        if m.get("online") and m.get("role") == "robot"]
-            if not machines:
-                log.write("[yellow]No online robots to test.[/]")
-                return
-
-            if not shutil.which("iperf3"):
-                log.write("[red]iperf3 not found locally — run: sudo apt install iperf3[/]")
-                return
-
-            server_ip = get_local_ip_for_network(topo, active_network)
-            if not server_ip:
-                log.write(f"[red]Cannot determine local IP on network '{active_network}'.[/]")
-                return
-
-            def on_results(results):
-                if not results:
-                    return
-                log.write(f"[dim]── bandwidth test summary ({active_network}) ──[/]")
-                for r in results:
-                    name = r.get("name", "?")
-                    if "error" in r:
-                        log.write(f"[red]{name}: {r['error']}[/]")
-                    else:
-                        mbps = r.get("mbps", 0)
-                        color = "green" if mbps > 10 else "yellow" if mbps > 1 else "red"
-                        tunnel_warn = "  [yellow]SSH tunnel (direct port blocked)[/]" if r.get("tunneled") else ""
-                        log.write(
-                            f"[{color}]{name}: {mbps:.1f} Mbps[/]"
-                            f"  retransmits={r.get('retransmits', 0)}"
-                            f"{tunnel_warn}"
-                        )
-
-            self.app.push_screen(BandwidthSelectScreen(machines, server_ip), on_results)
-
-    # ---- Zenoh Screen ----
-
-    class ZenohScreen(ModalScreen):
-        BINDINGS = [
-            Binding("escape", "dismiss", "Back", priority=True),
-            Binding("d", "deploy_all", "Deploy All", priority=True),
-            Binding("p", "preview", "Preview", priority=True),
-        ]
-
-        def compose(self) -> ComposeResult:
-            yield Vertical(
-                Label("[bold]Zenoh Config Manager[/]"),
-                Label(f"[dim]Network: {active_network} — Base stations connect to robots (star topology)[/]"),
-                Rule(),
-                DataTable(id="zenoh_table"),
-                Rule(),
-                RichLog(id="zenoh_log", highlight=True, markup=True, max_lines=30),
-            )
-            yield Footer()
-
-        def on_mount(self):
-            self._endpoints = {}  # machine_name -> list of endpoint strings
-            table = self.query_one("#zenoh_table", DataTable)
-            table.add_column("Machine", key="machine")
-            table.add_column("Role", key="role")
-            table.add_column("Connect Endpoints", key="connect")
-            table.add_column("Status", key="status")
-            table.cursor_type = "row"
-            self._build_table()
-
-        def _build_table(self):
-            table = self.query_one("#zenoh_table", DataTable)
-            table.clear()
-            log = self.query_one("#zenoh_log", RichLog)
-            self._endpoints = {}
-
-            for mname in sorted(runtime_config):
-                mconf = runtime_config[mname]
-                try:
-                    endpoints = generate_zenoh_endpoints(topo, active_network, mname)
-                    self._endpoints[mname] = endpoints
-                except KeyError:
-                    table.add_row(mname, mconf["role"], "?", "No address", key=mname)
-                    continue
-
-                n = len(endpoints)
-                connect_str = f"{n} robot(s)" if n else "none (listen only)"
-                table.add_row(mname, mconf["role"], connect_str, "Ready", key=mname)
-
-            log.write(f"[dim]Computed endpoints for {len(self._endpoints)} machine(s). Press [bold]d[/] to deploy, [bold]p[/] to preview.[/]")
-
-        def action_preview(self):
-            table = self.query_one("#zenoh_table", DataTable)
-            log = self.query_one("#zenoh_log", RichLog)
-            try:
-                from textual.coordinate import Coordinate
-                row_idx = table.cursor_coordinate.row
-                mname = str(table.get_cell_at(Coordinate(row_idx, 0)))
-            except Exception:
-                log.write("[red]No machine selected.[/]")
-                return
-
-            endpoints = self._endpoints.get(mname)
-            if endpoints is None:
-                log.write(f"[red]No config for {mname}.[/]")
-                return
-
-            log.write(f"\n[bold]Connect endpoints for {mname}:[/]")
-            if endpoints:
-                for ep in endpoints:
-                    log.write(f"  {ep}")
-            else:
-                log.write("  (none — listen only)")
-
-        def action_deploy_all(self):
-            log = self.query_one("#zenoh_log", RichLog)
-            if not self._endpoints:
-                log.write("[red]No machines to deploy to.[/]")
-                return
-
-            log.write("[dim]Deploying configs (updating connect.endpoints in existing JSON5)...[/]")
-
-            def do_deploy():
-                results = []
-                machines_to_deploy = []
-                for mname, endpoints in self._endpoints.items():
-                    mconf = runtime_config.get(mname)
-                    if not mconf or not mconf.get("online"):
-                        results.append((mname, False, "Offline"))
-                        continue
-                    machines_to_deploy.append((mname, mconf, endpoints))
-
-                def deploy_one(item):
-                    mname, mconf, endpoints = item
-                    ok, msg = deploy_zenoh_config(mconf["user"], mconf["ip"], endpoints)
-                    return mname, ok, msg
-
-                parallel_results = run_parallel(deploy_one, machines_to_deploy)
-                for item, result in parallel_results:
-                    if isinstance(result, Exception):
-                        results.append((item[0], False, str(result)))
-                    else:
-                        results.append(result)
-
-                return results
-
-            def on_done(results):
-                table = self.query_one("#zenoh_table", DataTable)
-                for mname, ok, msg in sorted(results):
-                    status = f"[green]{msg}[/]" if ok else f"[red]{msg}[/]"
-                    log.write(f"  {mname}: {status}")
-                    # Update table status column
-                    try:
-                        table.update_cell(mname, "status", msg)
-                    except Exception:
-                        pass
-                deployed = sum(1 for _, ok, _ in results if ok)
-                log.write(f"[bold]Deployed {deployed}/{len(results)} configs.[/]")
-
-            def bg():
-                results = do_deploy()
-                self.app.call_from_thread(on_done, results)
-
-            threading.Thread(target=bg, daemon=True).start()
-
-    # ---- Main App ----
-
-    class FleetApp(App):
-        CSS = """
-        Screen {
-            background: $surface;
-        }
-        #status_table {
-            height: auto;
-            max-height: 50%;
-        }
-        #log {
-            height: auto;
-            min-height: 5;
-        }
-        ModalScreen {
-            align: center middle;
-        }
-        """
-
-        BINDINGS = [
-            Binding("s", "refresh_status", "Recheck"),
-            Binding("l", "show_launch", "Launch"),
-            Binding("b", "show_maps", "Browse Maps"),
-            Binding("c", "show_config", "Config"),
-            Binding("o", "show_monitor", "Monitor"),
-            Binding("z", "show_zenoh", "Zenoh"),
-            Binding("q", "quit", "Quit"),
-        ]
-
-        def compose(self) -> ComposeResult:
-            yield Header()
-            yield Vertical(
-                Label(f"[bold]fleet-adt4[/]  network: [cyan]{active_network}[/]"),
-                Rule(),
-                DataTable(id="status_table"),
-                Rule(),
-                RichLog(id="log", highlight=True, markup=True, max_lines=100),
-            )
-            yield Footer()
-
-        def on_mount(self):
-            self.title = "fleet-adt4"
-            self.sub_title = f"network: {active_network}"
-            table = self.query_one("#status_table", DataTable)
-            table.add_column("Machine", key="machine")
-            table.add_column("Role", key="role")
-            table.add_column("Platform", key="platform")
-            table.add_column("IP", key="ip")
-            table.add_column("Ping", key="ping")
-            table.add_column("SSH", key="ssh")
-            table.add_column("Status", key="status")
-            self.action_refresh_status()
-
-        def action_refresh_status(self):
-            log = self.query_one("#log", RichLog)
-            log.write("[dim]Checking fleet connectivity (ping + SSH)...[/]")
-
-            def do_check():
-                machines = list(runtime_config.values())
-                reachable, unreachable = filter_reachable(machines)
-                return reachable + unreachable
-
-            def on_done(all_machines):
-                table = self.query_one("#status_table", DataTable)
-                table.clear()
-                n_online = 0
-                for m in sorted(all_machines, key=lambda x: x["name"]):
-                    ping_str = "OK" if m.get("ping") else "FAIL"
-                    ssh_str = "OK" if m.get("ssh") else "FAIL"
-                    if m.get("online"):
-                        status_str = "ONLINE"
-                        n_online += 1
-                    elif m.get("ping"):
-                        status_str = "NO SSH"
-                    else:
-                        status_str = "OFFLINE"
-                    table.add_row(
-                        m["name"],
-                        m["role"],
-                        runtime_config[m["name"]].get("platform_id", "---"),
-                        m["ip"],
-                        ping_str,
-                        ssh_str,
-                        status_str,
-                    )
-                    # Update runtime_config with reachability
-                    runtime_config[m["name"]].update({
-                        "ping": m.get("ping", False),
-                        "ssh": m.get("ssh", False),
-                        "online": m.get("online", False),
-                    })
-                log.write(f"[green]{n_online}/{len(all_machines)} machines online.[/]")
-
-            def bg():
-                results = do_check()
-                self.app.call_from_thread(on_done, results)
-
-            log = self.query_one("#log", RichLog)
-            threading.Thread(target=bg, daemon=True).start()
-
-        def action_show_maps(self):
-            self.push_screen(MapsScreen())
-
-        def action_show_launch(self):
-            # Create fresh each time — Textual doesn't preserve select()
-            # state across dismiss/re-push. The mode is re-applied on mount.
-            mode = getattr(self, "_last_launch_mode", None)
-            self.push_screen(LaunchScreen(mode=mode))
-
-        def action_show_config(self):
-            self.push_screen(ConfigScreen())
-
-        def action_show_monitor(self):
-            self.push_screen(MonitorScreen())
-
-        def action_show_zenoh(self):
-            self.push_screen(ZenohScreen())
-
-    # Auto-start zenohd if ROS 2 + rmw_zenoh_cpp are available and not already running.
-    # Uses the bundled zenoh_router_template.json5 (copied from a fleet machine) as the
-    # base config — preserves all ROS transport/timeout/routing settings. Patches only
-    # connect.endpoints and listen port from the topology.
-    _fleet_zenoh = {"proc": None, "config": None}  # mutable so closures can modify
-
-    def _stop_fleet_zenohd():
-        """Terminate the fleet-managed zenohd so the base station session can take over."""
-        proc = _fleet_zenoh["proc"]
-        if proc and proc.poll() is None:
-            proc.terminate()
-            try:
-                proc.wait(timeout=5)
-            except subprocess.TimeoutExpired:
-                proc.kill()
-                proc.wait(timeout=2)
-            import time
-            time.sleep(1)  # allow port to be released (TIME_WAIT)
-            _fleet_zenoh["proc"] = None
-
-    def _start_fleet_zenohd():
-        """Start (or restart) the fleet-managed zenohd on the standard zenoh port.
-
-        Returns (True, message) if started, (False, reason) otherwise.
-        """
-        if not shutil.which("ros2"):
-            return False, "ros2 not found on PATH — source your ROS workspace first"
-
-        # Don't start if one is already running under our control
-        if _fleet_zenoh["proc"] and _fleet_zenoh["proc"].poll() is None:
-            return True, "already running"
-
-        try:
-            local_ips = _local_ips
-            zenoh_conf = topo.get("zenoh", {})
-            port = zenoh_conf.get("port", 7447)
-
-            # Determine this machine's role from topology
-            local_machine = None
-            for m in runtime_config.values():
-                if m["ip"] in local_ips:
-                    local_machine = m
-                    break
-
-            # Build connect endpoints from topology
-            if local_machine:
-                endpoints = generate_zenoh_endpoints(
-                    topo, active_network, local_machine["name"]
-                )
-            else:
-                endpoints = [
-                    f"tcp/{m['ip']}:{port}"
-                    for m in runtime_config.values()
-                    if m.get("role") == "base_station" and m.get("ip")
-                ]
-
-            # Clean up previous config
-            if _fleet_zenoh["config"]:
-                try:
-                    os.unlink(_fleet_zenoh["config"])
-                except OSError:
-                    pass
-                _fleet_zenoh["config"] = None
-
-            config_path = patch_zenoh_config_local(endpoints)
-            if not config_path:
-                return False, "failed to patch zenoh config (template missing?)"
-
-            _fleet_zenoh["config"] = config_path
-            import tempfile
-            stderr_log = tempfile.NamedTemporaryFile(
-                mode="w", prefix="fleet_zenohd_", suffix=".log", delete=False
-            )
-            # Source the ROS workspace before running zenohd — fleet-adt4 may
-            # have been started without full ROS env (e.g. ros2 on PATH via
-            # .zshrc but workspace not sourced). Use zsh login shell so
-            # setup.zsh sources correctly and .zshrc env vars are available.
-            shell_cmd = (
-                f"source ~/dcist_ws/install/setup.zsh 2>/dev/null; "
-                f"ZENOH_ROUTER_CONFIG_URI={shlex.quote(config_path)} "
-                f"ros2 run rmw_zenoh_cpp rmw_zenohd"
-            )
-            zsh = shutil.which("zsh") or "zsh"
-            _fleet_zenoh["proc"] = subprocess.Popen(
-                [zsh, "-l", "-c", shell_cmd],
-                stdout=subprocess.DEVNULL,
-                stderr=stderr_log,
-            )
-            # Check it didn't die immediately
-            import time
-            time.sleep(1.0)
-            if _fleet_zenoh["proc"].poll() is not None:
-                rc = _fleet_zenoh["proc"].returncode
-                _fleet_zenoh["proc"] = None
-                stderr_log.close()
-                try:
-                    with open(stderr_log.name) as f:
-                        err_output = f.read().strip()[-500:]  # last 500 chars
-                except Exception:
-                    err_output = ""
-                detail = f"\n  stderr: {err_output}" if err_output else ""
-                return False, f"rmw_zenohd exited immediately (rc={rc}){detail}\n  config: {config_path}"
-            stderr_log.close()
-            return True, f"started on port {port} with {len(endpoints)} endpoint(s)"
-        except Exception as e:
-            return False, f"error: {e}"
-
-    try:
-        app = FleetApp()
-        app.run()
-    finally:
-        _stop_fleet_zenohd()
-        if _fleet_zenoh["config"]:
-            try:
-                os.unlink(_fleet_zenoh["config"])
-            except OSError:
-                pass
+from dcist_launch_system.tui.app import _launch_tui  # noqa: F401
 
 
 # ---------------------------------------------------------------------------
